[interpreter] Merge module-interpreter and expression interpreter
## Summary

Follow the convention in the same directory to have one interpreter for each program type.

## Test Plan

`yarn test`
